move exclusive caceh related code to exclusive
