+++ cache/msi.hpp
@@ -1,8 +1,6 @@
 #ifndef CM_CACHE_MSI_HPP
 #define CM_CACHE_MSI_HPP
 
-#include "cache/exclusive.hpp"
-
 // metadata supporting MSI coherency
 template <typename BT>
   requires C_SAME(BT, MetadataBroadcastBase) || C_SAME(BT, MetadataDirectoryBase)
@@ -110,38 +108,6 @@ public:
       return std::make_pair(false, cmd_for_null());
   }
 
-};
-
-template<typename MT, bool isLLC> requires C_DERIVE(MT, MetadataBroadcastBase)
-class ExclusiveMSIPolicy : public MSIPolicy<MT, false, isLLC>, public ExclusivePolicySupportBase    // always not L1
-{
-public:
-  virtual void meta_after_release(coh_cmd_t cmd, CMMetadataBase *mmeta, CMMetadataBase* meta, uint64_t addr, bool dirty){
-    // meta transfer from directory (if use directory coherence protocol) to cache
-    mmeta->init(addr);
-    mmeta->to_shared(-1); 
-    if(meta) { 
-      static_cast<MT*>(mmeta)->set_sharer(static_cast<MT*>(meta)->get_sharer());  
-      assert(!meta->is_dirty());
-      meta->to_invalid(); 
-    }
-    if(dirty) mmeta->to_dirty();
-  } 
-  virtual std::pair<bool, coh_cmd_t> release_need_probe(coh_cmd_t cmd, CMMetadataBase* meta) {
-    assert(this->is_release(cmd));
-    return std::make_pair(true, coh_cmd_t{cmd.id, this->probe_msg, this->evict_act});
-  }
-
-  virtual bool need_writeback(const CMMetadataBase* meta) { return true; }
-
-  virtual std::pair<bool, coh_cmd_t> inner_need_release(){
-    return std::make_pair(true, this->cmd_for_release());
-  }
-
-  virtual void meta_after_probe_ack(coh_cmd_t cmd, CMMetadataBase *meta, int32_t inner_id) const{
-    assert(this->is_probe(cmd));
-    if(this->is_evict(cmd)) meta->sync(inner_id);
-  }
 };
 
 #endif
